Split server components from client component facades
