+++ templates/demo-store-neue/src/components/cards/index.server.ts
@@ -1,0 +1,1 @@
+export {CollectionCard} from './CollectionCard.server';
