+++ templates/demo-store-neue/src/components/cards/index.ts
@@ -1,4 +1,3 @@
 export {ArticleCard} from './ArticleCard';
-export {CollectionCard} from './CollectionCard.server';
 export {OrderCard} from './OrderCard.client';
 export {ProductCard} from './ProductCard.client';
