+++ templates/demo-store-neue/src/components/global/index.server.ts
@@ -1,0 +1,2 @@
+export {NotFound} from './NotFound.server';
+export {Layout} from './Layout.server';
