+++ templates/demo-store-neue/src/components/global/index.ts
@@ -1,7 +1,5 @@
 export {Drawer, useDrawer} from './Drawer.client';
 export {Footer} from './Footer.client';
 export {Header} from './Header.client';
-export {Layout} from './Layout.server';
 export {Modal} from './Modal.client';
-export {NotFound} from './NotFound.server';
 export {PageHeader} from './PageHeader';
