+++ templates/demo-store-neue/src/components/index.server.ts
@@ -1,0 +1,5 @@
+export * from './cards/index.server';
+export * from './global/index.server';
+export * from './sections/index.server';
+export {CartProviderWithSession} from './CartProviderWithSession.server';
+export {DefaultSeo} from './DefaultSeo.server';
