+++ templates/demo-store-neue/src/components/index.ts
@@ -5,8 +5,6 @@ export * from './elements/index';
 export * from './global/index';
 export * from './product/index';
 export * from './sections/index';
-export {CartProviderWithSession} from './CartProviderWithSession.server';
 export {CountrySelector} from './CountrySelector.client';
 export {CustomFont} from './CustomFont.client';
-export {DefaultSeo} from './DefaultSeo.server';
 export {HeaderFallback} from './HeaderFallback';
