+++ templates/demo-store-neue/src/components/sections/index.server.ts
@@ -1,0 +1,1 @@
+export {ProductSwimlane} from './ProductSwimlane.server';
