+++ templates/demo-store-neue/src/components/sections/index.ts
@@ -1,3 +1,2 @@
 export {FeaturedCollections} from './FeaturedCollections';
 export {Hero} from './Hero';
-export {ProductSwimlane} from './ProductSwimlane.server';
